+++ lib/rootview.dart
@@ -19,7 +19,7 @@ class Page<T> {
   }
 }
 
-class NavigationBar extends StatelessWidget {
+class RootView extends StatelessWidget {
 
   static const String routeName = '/cupertino/navigation';
 
@@ -73,30 +73,4 @@ class NavigationBar extends StatelessWidget {
   }
 
   Future<bool> preventSwipePopping() => new Future<bool>.value(true);
-}
-
-class SettingsButton extends StatelessWidget {
-  const SettingsButton();
-
-  @override
-  Widget build(BuildContext context) {
-    return new CupertinoButton(
-      padding: EdgeInsets.zero,
-      child: new Image.asset(
-        'assets/ic_settings.png'), onPressed: () {},
-    );
-  }
-}
-
-class SearchButton extends StatelessWidget {
-  const SearchButton();
-
-  @override
-  Widget build(BuildContext context) {
-    return new CupertinoButton(
-      padding: EdgeInsets.zero,
-      child: new Image.asset(
-          'assets/ic_search.png'), onPressed: () {},
-    );
-  }
 }
